Router refactor done with auto drop route
